requirements.txt
Flask==3.0.0
Pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask-cors==4.0.0
rq==1.13.0
redis==4.7.0
flask-limiter==3.0.1